+++ src/main/java/kr/kakaocloud/kakeulgae/config/S3Config.java
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class TestS3Config {
+public class S3Config {
 
     @Value("${cloud.aws.credentials.access-key}")
     private String accessKey;
