+++ src/main/java/kr/kakaocloud/kakeulgae/controller/TestController.java
@@ -4,14 +4,19 @@ import kr.kakaocloud.kakeulgae.domain.dto.TestPostRequest;
 import kr.kakaocloud.kakeulgae.domain.dto.TestResponse;
 import kr.kakaocloud.kakeulgae.service.TestService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/test")
 public class TestController {
     private final TestService testService;
+
+    private final TestService imageService;
 
     @PostMapping()
     public void createTest(@RequestBody TestPostRequest t){
@@ -20,5 +25,11 @@ public class TestController {
     @GetMapping("{id}")
     public ResponseEntity<TestResponse> getName(@PathVariable(value="id") Long id) {
         return ResponseEntity.ok(testService.getId(id));
+    }
+
+    @PostMapping("/image")
+    @ResponseStatus(HttpStatus.OK)
+    public List<String> saveImage(@ModelAttribute TestPostRequest imageSaveDto) {
+        return imageService.saveImages(imageSaveDto);
     }
 }
