--- src/main/java/kr/kakaocloud/kakeulgae/domain/dto/TestImageSaveDto.java
@@ -1,17 +1,0 @@
-package kr.kakaocloud.kakeulgae.domain.dto;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class TestImageSaveDto {
-    private List<MultipartFile> images = new ArrayList<>();
-}
